refactor: :recycle: formatting and remove unused code
