+++ src/modules/accounts/infra/prisma/repositories/UsersRepository.ts
@@ -1,8 +1,8 @@
-import { PrismaClient } from '@prisma/client';
-import { getPrismaClient } from '@shared/container/providers/prisma';
-import { IUsersRepository } from '@modules/accounts/repositories/IUsersRepository';
 import { ICreateUser, IUpdateUser } from '@modules/accounts/dtos/Users';
 import User from '@modules/accounts/entities/User';
+import { IUsersRepository } from '@modules/accounts/repositories/IUsersRepository';
+import { PrismaClient } from '@prisma/client';
+import { getPrismaClient } from '@shared/infra/prisma';
 
 class UsersRepository implements IUsersRepository {
 	private prismaClient: PrismaClient;
