+++ src/modules/accounts/infra/prisma/repositories/UsersTokensRepository.ts
@@ -1,8 +1,8 @@
-import { PrismaClient } from '@prisma/client';
-import { getPrismaClient } from '@shared/container/providers/prisma';
-import { IUsersTokensRepository } from '@modules/accounts/repositories/IUsersTokensRepository';
 import { ICreateUserToken } from '@modules/accounts/dtos/UsersTokens';
 import UserToken from '@modules/accounts/entities/UserToken';
+import { IUsersTokensRepository } from '@modules/accounts/repositories/IUsersTokensRepository';
+import { PrismaClient } from '@prisma/client';
+import { getPrismaClient } from '@shared/infra/prisma';
 
 class UsersTokensRepository implements IUsersTokensRepository {
 	private prismaClient: PrismaClient;
